Create script2
This is the second script in the series. Again, straight into a "run applescript" and it should be good to go. All variables were set in the previous script so you the only change on this one is the the second repeat command should be set to the amount of sites being updated.